pyproject.toml
[project]
name = "til-application"
version = "0.1.0"
description = ""
authors = [
    {name = "Myeong Su Choi",email = "justdo.it.msc@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "mysqlclient (>=2.2.8,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
